refactor product-support logic to product and group plugins
