+++ services/web/server/src/simcore_service_webserver/groups/api.py
@@ -9,6 +9,7 @@ from ._groups_service import (
     get_product_group_for_user,
     get_user_profile_groups,
     is_user_by_email_in_group,
+    is_user_in_group,
     list_all_user_groups_ids,
     list_group_members,
     list_user_groups_ids_with_read_access,
@@ -23,6 +24,7 @@ __all__: tuple[str, ...] = (
     "get_product_group_for_user",
     "get_user_profile_groups",
     "is_user_by_email_in_group",
+    "is_user_in_group",
     "list_all_user_groups_ids",
     "list_group_members",
     "list_user_groups_ids_with_read_access",
