+++ services/web/server/src/simcore_service_webserver/products/products_web.py
@@ -3,6 +3,7 @@ from ._web_helpers import (
     get_current_product_credit_price_info,
     get_product_name,
     get_product_template_path,
+    is_user_in_product_support_group,
 )
 
 __all__: tuple[str, ...] = (
@@ -10,5 +11,6 @@ __all__: tuple[str, ...] = (
     "get_current_product_credit_price_info",
     "get_product_name",
     "get_product_template_path",
+    "is_user_in_product_support_group",
 )
 # nopycln: file
